Bug fixes for Mac OS build.
Minor bug and build fixes to get the Mac OS build fired up again.

+++ apps/ck/modules/json/encode.c
@@ -1079,7 +1079,7 @@ Return Value:
     Length = snprintf(Encoder->Result + Encoder->Length,
                       Encoder->Capacity - Encoder->Length,
                       "\n%*s",
-                      Count,
+                      (int)Count,
                       "");
 
     if (Length > 0) {
